pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cbir_feature_extractor"
version = "0.1.0"
description = "Módulo Python para extração de características de imagens testes de CBIR"
authors = [
    { name = "João Henrique V. do Carmo", email = "joaohvc1@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "kaggle>=1.6.0",
    "python-dotenv>=1.0.0"
]

# Configuração para que o setuptools encontre o seu código 
# dentro da pasta src/python automaticamente
[tool.setuptools.packages.find]
where = ["src"]
include = ["python*"]